requirements.txt
llama-index==0.10.68
llama-index-readers-file==0.1.33
llama-index-core==0.10.68
python-docx==1.1.2
trafilatura==1.12.0
beautifulsoup4==4.12.3
requests==2.32.3
chromadb==0.5.23
httpx==0.27.2
rank-bm25==0.2.2
sentence-transformers==3.3.1
python-dotenv==1.0.1
pydantic==2.9.2
loguru==0.7.2
tqdm==4.67.0
pytest==8.3.3
pytest-asyncio==0.24.0